refactor: Remove file_record structure and related functionality
Deleted the file_record structure and its associated methods from the
codebase. Updated file_index_manager and base_event_handler to remove
dependencies on file_record. Introduced a new inline function to create
file records directly within file_index_manager, enhancing
encapsulation and reducing complexity.

Log: Refactoring source code
