+++ src/server/include/core/file_index_manager.h
@@ -14,7 +14,6 @@
 #include <lucene++/LuceneHeaders.h>
 
 #include "common/anything_fwd.hpp"
-#include "common/file_record.h"
 #include "core/pinyin_processor.h"
 
 ANYTHING_NAMESPACE_BEGIN
@@ -46,8 +45,6 @@ public:
     /// @param exact_match If true, performs an exact match when removing the old path; 
     ///                    otherwise, performs a fuzzy match.
     void update_index(const std::string& old_path, const std::string& new_path, bool exact_match = true);
-
-    // std::vector<file_record> search_index(const std::string& term, bool exact_match = false, bool nrt = false);
 
     /// Commit all changes to the index
     void commit(index_status status);
@@ -98,21 +95,6 @@ public:
     ///         If no files are found, an empty list is returned.
     QStringList search(const QString& path, QString& keywords, const QString& type, bool nrt);
 
-    /// @brief Searches all files created between a specified time range, returning a list of matching file names.
-    /// @param keywords The keywords to search for.
-    /// @param after The start time for the search range. Files created after this time will be included in the search.
-    ///              The time should be provided in a recognized format (e.g., "YYYY-MM-DD HH:MM:SS"). If this parameter is empty,
-    ///              the search will include all files created before the `before` parameter.
-    /// @param before The end time for the search range. Files created before this time will be included in the search.
-    ///               The time should be provided in a recognized format (e.g., "YYYY-MM-DD HH:MM:SS"). If this parameter is empty,
-    ///               the search will include all files created after the `after` parameter.
-    ///               If both `after` and `before` are empty, the search will include all files matching the keywords regardless of time.
-    /// @param nrt If true, performs a near real-time search that includes recent changes not yet committed; 
-    ///            otherwise, searches the last committed index for more stable results.
-    /// @return A QStringList containing the paths of all files where the keywords are found within the specified time range. 
-    ///         If no files are found, an empty list is returned.
-    QStringList search(const QString& path, QString& keywords, const QString& after, const QString& before, bool nrt);
-
     void async_search(QString& keywords, bool nrt, std::function<void(const QStringList&)> callback);
 
     QStringList traverse_directory(const QString& path, bool nrt);
@@ -144,8 +126,6 @@ private:
      */
     // Lucene::TopScoreDocCollectorPtr nrt_search(const std::string& path, bool exact_match = false);
 
-    Lucene::DocumentPtr create_document(const file_record& record);
-
     void prepare_index();
 
     void check_index_version();
@@ -170,7 +150,6 @@ private:
     Lucene::String pinyin_field_{ L"pinyin" };
     std::mutex mtx_;
     std::mutex reader_mtx_;
-    file_helper file_helper_;
     pinyin_processor pinyin_processor_;
     std::atomic<bool> search_cancelled_{false};
 };
